--- a/export_test.go
+++ b/export_test.go
@@ -21,3 +21,15 @@ func CollectMapStats(m *Map) MapStats {
 func MapSize(m *Map) int {
 	return m.size()
 }
+
+func TopHashMatch(hash, topHashes uint64, idx int) bool {
+	return topHashMatch(hash, topHashes, idx)
+}
+
+func StoreTopHash(hash, topHashes uint64, idx int) uint64 {
+	return storeTopHash(hash, topHashes, idx)
+}
+
+func EraseTopHash(topHashes uint64, idx int) uint64 {
+	return eraseTopHash(topHashes, idx)
+}